+++ App/App.tsx
@@ -1,6 +1,7 @@
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { LoginPage } from './screens/LoginPage';
+import { Login } from './screens/Login';
+import { Signup } from './screens/Signup';
 import React from 'react';
 
 const Stack = createNativeStackNavigator();
@@ -9,7 +10,8 @@ export default function App() {
   return (
     <NavigationContainer>
       <Stack.Navigator>
-        <Stack.Screen name="Login" component={LoginPage} />
+        <Stack.Screen name="Signup" component={Signup} />
+        <Stack.Screen name="Login" component={Login} />
       </Stack.Navigator>
     </NavigationContainer>
   );
